Remove SimpleQueryEngine
